Improve the art style, AI, add more furniture

+++ src/shared/network/event/SoundEvent.java
@@ -9,9 +9,9 @@ public class SoundEvent extends GameEvent {
 	public static final int PING_SOUND_ID = 34;
 	public static final int DOOR_OPEN_ID = 36;
 	
-	public static final double BULLET_WALL_SOUND_VOLUME = 25;
-	public static final double PING_SOUND_VOLUME = 30;
-	public static final double FOOTSTEP_SOUND_VOLUME = 15;
+	public static final float BULLET_WALL_SOUND_VOLUME = 25;
+	public static final float PING_SOUND_VOLUME = 30;
+	public static final float FOOTSTEP_SOUND_VOLUME = 15;
 	
 	public final float x;
 	public final float y;
